+++ NEWS.md
@@ -69,6 +69,7 @@ Note: We're only listing outstanding class updates.
 
 * Proc
     * Proc#dup returns an instance of subclass. [[Bug #17545]]
+    * Proc#parameters now accepts lambda keyword. [[Feature #15357]]
 
 * Refinement
     * Refinement#refined_class has been added. [[Feature #12737]]
@@ -81,6 +82,7 @@ Note: We're only listing outstanding class updates.
     * String#byteindex and String#byterindex have been added. [[Feature #13110]]
     * Update Unicode to Version 14.0.0 and Emoji Version 14.0. [[Feature #18037]]
       (also applies to Regexp)
+    * String#bytesplice has been added.  [[Feature #18598]]
 
 * Struct
     * A Struct class can also be initialized with keyword arguments
@@ -170,6 +172,7 @@ The following deprecated APIs are removed.
 [Feature #13110]: https://bugs.ruby-lang.org/issues/13110
 [Feature #14332]: https://bugs.ruby-lang.org/issues/14332
 [Feature #15231]: https://bugs.ruby-lang.org/issues/15231
+[Feature #15357]: https://bugs.ruby-lang.org/issues/15357
 [Bug #15928]:     https://bugs.ruby-lang.org/issues/15928
 [Feature #16131]: https://bugs.ruby-lang.org/issues/16131
 [Feature #16806]: https://bugs.ruby-lang.org/issues/16806